Use subreddit comments api to additionally check for all recent comments
Creates a new collection in firestore for tracking the progress of
comment processing so that we eventually only check the latest
comments.

Uses the
[`comments`](https://praw.readthedocs.io/en/latest/code_overview/models/subreddit.html#praw.models.Subreddit.comments)
api to get the latest comments (by default fetches limits to 100) of
the subreddit. And tracks the earliest and latest comments it
found. Subsequent runs uses these "cursors" to fetch any newer and
older comments.